requirements.txt
black==24.10.0
isort==5.13.2
ruff==0.8.0
pre_commit==4.0.1
python-dotenv==1.0.1
requests==2.32.3
mortgage==1.0.5
pandas==2.2.3
yfinance==0.2.50
fastapi
uvicorn[standard]
playwright
pytest-playwright
